refactor: 將 App.vue 簡化為純根元件，業務邏輯移至 AppLayout
- 新建 AppLayout.vue，承接全局狀態、provide、getData 與 MainNavBar
- App.vue 僅保留 router-view 與全局 SCSS 樣式
- 路由重構：AppLayout 作為根 Layout，MenuPage 移至 /menu 路徑
- MealView 改為動態導向第一個分類，移除寫死的 /Beef
- InfoBox 改用 inject('diningHours') 取代 prop 傳遞
- CategoryNav 路由路徑更新為 /menu/:category
