+++ src/components/menu-page/CategoryNav.vue
@@ -2,7 +2,7 @@
   <ul class="nav flex-xl-column flex-nowrap overflow-x-auto overflow-y-hidden">
     <li v-for="category in categoryList" :key="category" class="nav-item">
       <RouterLink
-        :to="{ path: `/${category}` }"
+        :to="{ path: `/menu/${category}` }"
         class="nav-link fw-normal px-xl-5"
         :class="{ active: false }"
       >
