+++ src/components/menu-page/InfoBox.vue
@@ -16,8 +16,7 @@
 <script setup>
 import { inject } from 'vue'
 
-defineProps(['diningHours'])
-
+const diningHours = inject('diningHours')
 const tableNumber = inject('tableNumber')
 const guestsCount = inject('guestsCount')
 </script>
